docker 컨테이너중 멈춰 있는 컨테이너도 확인해 mark_cont 가 있으면 삭제

+++ docker_run_shell.sh
@@ -5,9 +5,9 @@ if [[ "$(docker images -q mark_img:0.1 2> /dev/null)" == "" ]]; then
   docker build -t mark_img:0.1 .
 fi
 
-if [[ "$(docker ps | grep mark_cont 2> /dev/null)" != "" ]]; then
+if [[ "$(docker ps -a | grep mark_cont 2> /dev/null)" != "" ]]; then
   docker rm -f mark_cont
 fi
 
-docker run -itd --name mark_cont -v ${DIR}/public:/data/public -p 81:3000 mark_img:0.1 bash
-docker exec mark_cont forever start /data/bin/www
+docker run -itd --name mark_cont -v ${DIR}/public:/data/public -p 81:3000 mark_img:0.1
+
